+++ src/template/js/src/server/controllers/assets.js
@@ -4,7 +4,18 @@ import path from "path";
 class AssetsController {
   router = express.Router();
 
-  static assetList = [];
+  static assetList = [
+    "android-chrome-192x192.png",
+    "android-chrome-512x512.png",
+    "apple-touch-icon.png",
+    "browserconfig.xml",
+    "favicon-16x16.png",
+    "favicon-32x32.png",
+    "favicon.ico",
+    "mstile-150x150.png",
+    "safari-pinned-tab.svg",
+    "site.webmanifest",
+  ];
 
   constructor() {
     this.initializeRoutes();